Modifications for the Akai APC 20.
The abstractions which handle channels were getting in my way, so I
simply disabled them. This apc20.rb script simply picks up incoming
MIDI packets and reports on the data they contain.

With this script, I have reliable access to 87 controllers on the APC:
54 punch buttons, 24 toggle buttons, and 9 sliders. The remaining
controllers (a dial and a few transport control buttons) give MIDI
packets which are too weird to be worth giving a damn about.

It's likely I can use this code to control Archaeopteryx in real time
in some pretty interesting ways, but I don't yet know for sure.